ranking_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
numpy
scikit-learn
pandas
nltk
python-docx
pdfminer.six
pandas 
xlsxwriter
